% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response-package.R
\docType{package}
\name{response-package}
\alias{response-package}
\alias{_PACKAGE}
\title{response: Mean and Confidence Intervals of a Response Variable by Group}
\description{
Gives mean and confidence intervals of a response variable by groups of an input variable. The response can be numeric with confidence intervals calculated using Gaussian assumptions, or if the response has two classes then a binomial c confidence interval can be used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jedwards24/response}
  \item Report bugs at \url{https://github.com/jedwards24/response/issues}
}

}
\author{
\strong{Maintainer}: James Edwards \email{jamese1326@gmail.com}

}
\keyword{internal}
