% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\title{Confidence intervals for binary target variable by values of a discrete predictor}
\usage{
response(
  dt,
  target_name,
  var_name,
  min_n = 1,
  show_all = TRUE,
  order_n = NULL,
  conf_level = 0.95,
  pos_class = NULL,
  plot = TRUE,
  family = "guess"
)
}
\arguments{
\item{dt}{A data frame.}

\item{target_name}{Column to use as target variable. Column name (quoted or unquoted) or position.}

\item{var_name}{Column to use as predictor variable. Column name (quoted or unquoted) or position.}

\item{min_n}{Integer >= 1. Predictor levels with less than \code{min_n} observations are not displayed.}

\item{show_all}{Logical. Defaults to \code{TRUE}. If \code{FALSE} will not show levels whose confidence interval
overlaps the mean response of all observations.}

\item{order_n}{Logical. Whether to force plot and table to order by number of observations of the
predictor. The default setting \code{NULL} retains ordering if predictor is numeric or ordered factor
and orders by number of observations otherwise.}

\item{conf_level}{Numeric in (0,1). Confidence level used for confidence intervals.}

\item{pos_class}{Optional. Specify value in target to associate with class 1.}

\item{plot}{Logical. If \code{TRUE}, a plot using \code{plot_reponse()} will be printed.}

\item{family}{Distribution family to use for the response confidence interval. Either "binomial"
or "gaussian".}
}
\value{
A summary data frame. By default a plot will also be printed as a side effect.
}
\description{
In the \code{data}, one column is a response variable (\code{target_name}) and another is a predictor or
grouping variable (\code{var_name}). A mean and confidence interval of the response for each value
of the predictor is calculated and returned in a data frame. The function most appropriate for
factor predictors where grouping is more natural but will also work with other variable types.
}
\details{
The target variable must be binary. To compute confidence intervals, values are treated as 0
and 1 values. Which value corresponds to 1 (the "positive class") is controlled by the argument
\code{pos_class}. If not supplied, the positive class will be the "greatest" value using natural
ordering (level ordering if target is an ordered factor).
}
